Require screen recording permissions on Catalina, update copyrights
macOS Catalina requires a new permission to read the window title from
window enumeration APIs. The symptom of the problem of running without
permissions is that the kCGWindowName field in the dictionaries returned
by CGWindowListCopyWindowInfo() is not populated.

Update winutils (commit 9227e332 of https://github.com/andrewgho/movewin.git)
to add a function to check for screen recording permissions by empirically
trying to start a screen recording session (without actually doing any
recording), and update the Ruby code to match.

Also, update copyrights to the current year, which is the best of all years;
and fix some broken include clauses in the example programs.